requirements.txt
streamlit
python-docx
pdfminer.six
spacy
scikit-learn
fuzzywuzzy
openai
python-dotenv
